# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_xilo-platform_xilo_internal_auth_model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  github_com_xilo-platform_xilo_internal_post_model.CreatePostRequest:
    properties:
      category:
        type: string
      content:
        type: string
      content_md:
        type: string
      cover_image_url:
        type: string
      excerpt:
        type: string
      is_premium:
        type: boolean
      scheduled_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_post_model.Post:
    properties:
      author:
        $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.User'
      author_id:
        type: string
      category:
        type: string
      content:
        type: string
      content_md:
        type: string
      cover_image_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      excerpt:
        type: string
      id:
        type: string
      is_premium:
        type: boolean
      published_at:
        type: string
      reading_time:
        type: integer
      scheduled_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      word_count:
        type: integer
    type: object
  github_com_xilo-platform_xilo_internal_post_model.UpdatePostRequest:
    properties:
      category:
        type: string
      content:
        type: string
      content_md:
        type: string
      cover_image_url:
        type: string
      excerpt:
        type: string
      is_premium:
        type: boolean
      scheduled_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@xilo.example.com
    name: Xilo API Support
    url: https://xilo.example.com/support
  description: A self-hosted, Telegram-inspired modern blog platform API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://xilo.example.com/terms
  title: Xilo Blog Platform API
  version: "1.0"
paths:
...
  /notifications:
    get:
      parameters:
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}/read:
    post:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /notifications/preferences:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get notification preferences
      tags:
      - notifications
    patch:
      consumes:
      - application/json
      parameters:
      - description: Preferences map
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update notification preferences
      tags:
      - notifications
  /notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/sms/broadcast:
    post:
      consumes:
      - application/json
      parameters:
      - description: Broadcast data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Broadcast SMS to all users
      tags:
      - notifications
  /notifications/sms/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: SMS data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send SMS notification
      tags:
      - notifications
  /posts:
    get:
      description: List published blog posts with pagination and filters
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Filter by author
        in: query
        name: author
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List published posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new blog post
      parameters:
      - description: Post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_post_model.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_post_model.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update an existing blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_post_model.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_post_model.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
...
  /posts/{slug}:
    get:
      description: Get a post by its slug
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_post_model.Post'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get post by slug
      tags:
      - posts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"